feat: Album View
Album View:
* 新增專輯頁面
* 新增專輯照片
* 新增歌曲列表

搜尋頁面:
* 在專輯欄目增加NavigationLink連結到相應的頁面

Song View:
* Bottom toolbar的專輯button導向相應的頁面

+++ iTunesSearch/DataModel/ArtistAlbumLookup.swift
@@ -27,7 +27,7 @@ struct ArtistAlbumLookup: Codable {
     let collectionName, collectionCensoredName: String?
     let artistViewURL, collectionViewURL: String?
     let artworkUrl60, artworkUrl100: String?
-    let collectionPrice: Int?
+    let collectionPrice: Double?
     let collectionExplicitness: String?
     let trackCount: Int?
     let copyright: String?
